test(WinningLotto): 쉼표로 구분된 숫자를 받아 배열로 반환하는 기능 테스트
